+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v7.4.1 - 2022-12-23
+
+Added
+  * GH-11: Add domain related outputs
+
+# v7.4.0 - 2022-12-16
+
+Added
+  * AZ-928: Add authentication and client certificates parameters
+
 # v7.3.0 - 2022-12-14
 
 Added